refactor: created aws package
