+++ extended/src/main/java/apoc/ml/aws/SageMakerConfig.java
@@ -1,4 +1,4 @@
-package apoc.ml.bedrock;
+package apoc.ml.aws;
 
 import java.util.Map;
