+++ extended/src/test/java/apoc/ml/aws/BedrockIT.java
@@ -1,4 +1,4 @@
-package apoc.ml.bedrock;
+package apoc.ml.aws;
 
 import apoc.util.TestUtil;
 import org.apache.commons.codec.binary.Base64;
@@ -15,13 +15,13 @@ import java.util.Map;
 import static apoc.ApocConfig.apocConfig;
 import static apoc.ExtendedApocConfig.APOC_AWS_KEY_ID;
 import static apoc.ExtendedApocConfig.APOC_AWS_SECRET_KEY;
-import static apoc.ml.bedrock.AWSConfig.KEY_ID;
-import static apoc.ml.bedrock.AWSConfig.METHOD_KEY;
-import static apoc.ml.bedrock.AWSConfig.SECRET_KEY;
-import static apoc.ml.bedrock.BedrockGetModelsConfig.*;
-import static apoc.ml.bedrock.BedrockInvokeConfig.MODEL;
-import static apoc.ml.bedrock.BedrockTestUtil.*;
-import static apoc.ml.bedrock.BedrockUtil.*;
+import static apoc.ml.aws.AWSConfig.KEY_ID;
+import static apoc.ml.aws.AWSConfig.METHOD_KEY;
+import static apoc.ml.aws.AWSConfig.SECRET_KEY;
+import static apoc.ml.aws.BedrockGetModelsConfig.*;
+import static apoc.ml.aws.BedrockInvokeConfig.MODEL;
+import static apoc.ml.aws.BedrockTestUtil.*;
+import static apoc.ml.aws.BedrockUtil.*;
 import static apoc.util.TestUtil.testCall;
 import static apoc.util.TestUtil.testResult;
 import static org.junit.Assert.assertEquals;
